+++ CommonModules/Utilities.py
@@ -2,6 +2,12 @@
 __author__ = "Wang Hewen"
 """ Some other functionalities."""
 import datetime
+import time
+
+global CurrentTime
+global StartTime
+CurrentTime = 0
+StartTime = 0
 
 def ConvertTimeStampToDateTime(TimeStamp, TimeType = 0, ReturnType = 0):
     '''
@@ -40,3 +46,33 @@ def ConcatenateIntegers(*Integers):
     for Integer in Integers:
         Result += str(Integer)
     return int(Result)
+
+def TimeElapsed(Unit = True, LastTime = False):
+    '''
+    Return the time interval since first/last call of this function.
+
+    :param Boolean/String Unit: Whether to append unit ' sec'(by default) or other unit when returned. If False, a float number will be returned.
+    :param Boolean LastTime: If True return the time interval since the last call of this function, otherwise return the time interval since the first call of this function, e.g. since programming running.
+    :return: TimeInterval: Time interval since last call of this function..
+    :rtype: String/Float
+    '''
+    global CurrentTime
+    global StartTime
+
+    if LastTime:
+        TimeInterval = time.time() - CurrentTime#Basically it's the previous time
+    else:
+        TimeInterval = time.time() - StartTime
+
+    if StartTime == 0:
+        TimeInterval = 0.0
+        StartTime = time.time()
+
+    CurrentTime = time.time()
+    
+    if Unit == True:
+        return str(TimeInterval) + " sec"
+    elif Unit:
+        return str(TimeInterval) + Unit
+    else:
+        return TimeInterval