+++ Test.py
@@ -1,14 +1,19 @@
  # -*- coding:utf-8 -*-
 __author__ = "Wang Hewen"
 import numpy as np
-import CommonModules.DataStructureOperations as DSO
-import CommonModules.IO as IO
+import CommonModules.DataStructureOperations
+import CommonModules.IO
+from CommonModules.Utilities import TimeElapsed
+
+CM = CommonModules
 
 def main():
+    print(TimeElapsed())
     test_array = np.array([[4,5,6]])
     #test_array = np.array([])
-    print(DSO.CombineSparseMatricesRowWise(test_array, np.array([[1,2,3]])).toarray())
-    IO.ExportNpArray("./test.txt", test_array)
+    print(CM.DataStructureOperations.CombineMatricesRowWise(test_array, np.array([[1,2,3]])))
+    #CM.IO.ExportNpArray("./test.txt", test_array)
+    print(TimeElapsed(LastTime = True))
     return
 
 if __name__ == "__main__":